add permission to route

+++ routes/frontend/coupon.php
@@ -13,32 +13,32 @@ Route::group(['as' => 'coupons.', 'prefix' => 'coupons', 'middleware' => ['auth'
         ->breadcrumbs(function (Trail $trail) {
             $trail->parent(homeRoute())
                 ->push(__('Coupon management'), route('frontend.coupons.index'));
-        });
+        })->middleware('permission:user.coupon.view');
 
     Route::get('create', [CouponController::class, 'create'])->name('create')
         ->breadcrumbs(function (Trail $trail) {
             $trail->parent(homeRoute())
                 ->push(__('Coupon management'), route('frontend.coupons.index'))
                 ->push(__('Create new coupon'));
-        });
+        })->middleware('permission:user.coupon.create');
 
-    Route::post('store', [CouponController::class, 'store'])->name('store');
+    Route::post('store', [CouponController::class, 'store'])->name('store')->middleware('permission:user.coupon.store');
 
     Route::get('{slug}/edit', [CouponController::class, 'edit'])->name('edit')
         ->breadcrumbs(function (Trail $trail) {
             $trail->parent(homeRoute())
                 ->push(__('Coupon management'), route('frontend.coupons.index'))
                 ->push(__('Update coupon'));
-        });
+        })->middleware('permission:user.coupon.edit');
 
     Route::get('{slug}/detail', [CouponController::class, 'detail'])->name('detail')
         ->breadcrumbs(function (Trail $trail) {
             $trail->parent(homeRoute())
                 ->push(__('Coupon management'), route('frontend.coupons.index'))
                 ->push(__('Coupon Information'));
-        });
+        })->middleware('permission:user.coupon.detail');
 
-    Route::put('{slug}/update', [CouponController::class, 'update'])->name('update');
+    Route::put('{slug}/update', [CouponController::class, 'update'])->name('update')->middleware('permission:user.coupon.edit');
 
-    Route::delete('{slug}/destroy', [CouponController::class, 'destroy'])->name('destroy');
+    Route::delete('{slug}/destroy', [CouponController::class, 'destroy'])->name('destroy')->middleware('permission:user.coupon.disable');
 });
